Moving redis commands to a mixin
This patterns allows for the reuse of these commands across connection types, including modules that are based on this client
